pyproject.toml
[tool.poetry]
name = "erpbackend"
version = "0.1.0"
description = ""
authors = ["Sreedeep CV <sreedeep200@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.7"
psycopg = "^3.1.8"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
coreapi = "^2.3.3"
django-cors-headers = "^3.14.0"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
django-rest-knox = "^4.2.0"
openpyxl = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
